test/robustness: Create dedicated traffic package
